requirements.txt
tensorflow>=2.10
pandas
numpy
scikit-learn
statsmodels
pmdarima
matplotlib
seaborn
streamlit
folium
streamlit-folium
pyyaml
xlrd
pytest
